hearts/requirements.txt
boto3
psycopg2-binary
firebase_admin
pytz
requests
